Fixed errors, refactored stat reset and added CharacterStatEnum

+++ src/Imgeneus.World/Game/Player/DefaultLevelStatSkillPoints.cs
@@ -5,7 +5,7 @@ namespace Imgeneus.World.Game.Player
     public class DefaultLevelStatSkillPoints
     {
         public Mode Mode { get; set; }
-        public byte StatPoint { get; set; }
-        public byte SkillPoint { get; set; }
+        public ushort StatPoint { get; set; }
+        public ushort SkillPoint { get; set; }
     }
 }
